pkgdown/extra.scss

Sample parameters to run population projection

Data frame containing the rates and number of people from the Federal Statistical Office (FSO) required to project the development of four demographic groups for a selected canton (Aargau). The parameters are from the model published in 2020. The sample data only include the reference scenario and the years 2019-2030.

Usage

fso_parameters

Format

The example data include the required parameters for each demographic group (nationality (2) X sex (2) X age classes (101)) for the years 2019-2030.

Source

Data obtained from the Swiss Federal Statistical Office (FSO):

https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_101/-/px-x-0104020000_101.px/

https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_102/-/px-x-0104020000_102.px

https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_103/-/px-x-0104020000_103.px/

https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_106/-/px-x-0104020000_106.px/

https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_109/-/px-x-0104020000_109.px/

Demographic groups

The returned data frame includes parameters for each unique combination of the following demographic variables:

nat: ch = Swiss; int = foreign / international.

sex: f = female, m = male.

age: 101 one-year age classes, ranging from 0 to 100 (including those older than 100).

Parameters

The following parameters are included in the returned data frame:

year: character, year of projection.

scen: character, projection scenario.

birth_rate: numeric, number of children per year.

births_int_ch: numeric, proportion of children with Swiss nationality born to non-Swiss mothers.

mor: numeric, prospective mortality (probability of death).

emi: numeric, rate of people emigrating abroad.

acq: numeric, rate of acquisition of Swiss citizenship.

imm_int: numeric, number of people immigrating from abroad.

mig_ch: numeric, national / inter-cantonal net migration (number of immigrants minus number of emigrants).

spatial_unit: character, indicating the user requested spatial unit(s).

Details about calculated variables

births_int_ch is calculated by dividing the number of live newborns with Swiss citizenship born to non-Swiss mothers by the number of all live newborns born to non-Swiss mothers.

mig_ch is calculated as the difference between the immigration from other cantons and the emigration to other cantons.