pkgdown/extra.scss

propop: Project Population Growth in Switzerland using the Cohort Component Method

The purpose of this package is to project the development of the population at different spatial levels (e.g., cantons, districts, municipalities) using the cohort component method and the parameters provided by the Federal Statistical Office (FSO).

See also

Author

Maintainer: Norah Efosa norah.efosa@ag.ch

Authors:

Adrian Gadient adrian.gadient@ag.ch (ORCID)

Tina Cornioley tina.cornioley@vd.ch

Jan Wunder jan.wunder@sg.ch

Niklas Haffert niklas.haffert@sg.ch

Other contributors:

Andrea Plüss andrea.pluess@ag.ch [contributor]

Nadine Herrmann [contributor]

Lutz Benson [contributor]

Thomas Knecht [contributor]

Eric Meyer [contributor]

Statistik Aargau [funder, copyright holder]