pkgdown/extra.scss

Project population development (raw results)

Core function that uses the cohort component method and matrix algebra to project population development. The function can be used for different spatial levels (e.g., cantons, municipalities) and for one scenario at a time.

This basic function provides projections in a raw version in which key information is missing (e.g., which age groups the rows represent). To conveniently obtain an enriched, more informative output, use the wrapper function propop::propop() (which internally uses propop::project_raw()).

The parameters and start populations for different spatial levels can be obtained from the Swiss Federal Statistical Office (FSO). For instructions on how to download this information from STAT-TAB, see vignette("prepare_data", package = "propop").

The projection parameters need to be passed on as a single data frame to project_raw with (with the parameters as columns). The column types, names, and factor levels need to match those specified below.

Usage

project_raw(
  parameters,
  year_first,
  year_last,
  age_groups = 101,
  fert_first = 16,
  fert_last = 50,
  share_born_female = 100/205,
  n,
  subregional
)

Arguments

parameters

data frame containing the FSO rates and numbers to run the projection for a specific spatial level (e.g., canton, municipality).

year: projection year.

spatial_unit: ID of spatial entity (e.g., canton, municipality) for which to run the projections.

scen: projection scenario, used to subset data frames with multiple scenarios (r = reference, l = low growth, h = high growth scenario).

nat: nationality (ch = Swiss; int = foreign / international).

sex: sex (f = female, m = male).

age: age classes; typically ranging from 0 to 100 (incl. >100).

birth_rate: number of children per year.

births_int_ch proportion of children with Swiss nationality born to non-Swiss mothers.

mor: prospective mortality rate (probability of death).

acq: rate of acquisition of Swiss citizenship.

emi: rate of people emigrating abroad.

mig_ch: national / inter-cantonal net migration (number of immigrants - number of emigrants).

imm_int: number of people immigrating from abroad.

mig_sub: within canton net migration. Useful to account for movements between different subregions (e.g., municipalities). This argument is optional.

year_first

numeric, first year to be projected.

year_last

numeric, last year to be projected.

age_groups

numeric, number of age classes. Creates a vector with 1-year age classes running from 0 to (age_groups - 1). Defaults to 101 (FSO standard number of age groups).

fert_first

numeric, first year of female fertility. Defaults to 16 (FSO standard value).

fert_last

numeric, last year of female fertility. Defaults to 50 (FSO standard value).

share_born_female

numeric, fraction of female babies. Defaults to 100 / 205 (FSO standard value).

n

number of people per demographic group and year; should be the year before year_first. Typically extracted from data frame created with propop::get_population().

subregional

boolean, TRUE indicates that subregional migration patterns (e.g., movement between municipalities within a canton) are part of the projection.

Value

Returns an unformatted and unlabeled data frame. It includes the number of people for each demographic group per year (starting year and projected years. The number of rows corresponds to the product of years and demographic groups (e.g., nationality (2) X sex (2) X age groups (101) = 404). Variables included in the output:

n

number of people per demographic group.

IMM_INT

number of immigrants from other countries.

MIG_CH

number of people migrating from / to other superordinate spatial units (typically cantons).

MIG_SUB

number of migrants within the superordinate spatial unit (typically a canton).

MOR

number of deaths (among people older than 0).

EMI_INT

number of emigrants to other countries.

ACQ

number of foreigners who acquire Swiss citizenship (naturalisations).

BIRTHS

number of births.

See also

Examples

# load package data
data(fso_parameters)
data(fso_population)

# run projection
project_raw(
  parameters = fso_parameters,
  year_first = 2019,
  year_last = 2019,
  n = fso_population |> dplyr::pull(n),
  subregional = FALSE
) |>
  head(10)
#> Running projection for: Aargau
#> ✔ Year: 2019
#>       N IMM_INT MIG_CH MIG_SUB MOR EMI_INT ACQ BIRTHS
#> 1  2506       2     17       0  NA      NA  NA     NA
#> 2  2589      15     34       0  NA      NA  NA     NA
#> 3  2568      15     26       0  NA      NA  NA     NA
#> 4  2635      14     23       0  NA      NA  NA     NA
#> 5  2710      13     16       0  NA      NA  NA     NA
#> 6  2608      11     10       0  NA      NA  NA     NA
#> 7  2569      10      7       0  NA      NA  NA     NA
#> 8  2611       9      6       0  NA      NA  NA     NA
#> 9  2634       8      2       0  NA      NA  NA     NA
#> 10 2512       7      1       0  NA      NA  NA     NA