pkgdown/extra.scss

Project population development (enriched results)

Project population development using the cohort component method (see e.g., here for more details). This wrapper function runs the projection and returns a clean data frame with clearly labeled variables and variable levels.

The parameters and start population can be obtained from the Swiss Federal Statistical Office (FSO). For instructions on how to download this information from STAT-TAB, see vignette("prepare_data", package = "propop").

For more details on how to use this function to project the population development on the level of a canton, see vignette("run_projections", package = "propop").

The projection parameters need to be passed to propop::propop() as a single data frame (with the parameters as columns). The column types, names, and factor levels need to match the specifications listed below under parameters:

Usage

propop(
  parameters,
  population,
  year_first,
  year_last,
  age_groups = 101,
  fert_first = 16,
  fert_last = 50,
  share_born_female = 100/205,
  subregional,
  spatial_unit = "spatial_unit"
)

Arguments

parameters

data frame containing the FSO rates and numbers to run the projection for a specific spatial level (e.g., canton, municipality).

year character, projection year.

spatial_unit, character, ID of spatial entity (e.g., canton, municipality) for which to run the projections.

scen, character, projection scenario, is used to subset data frames with multiple scenarios (r = reference, l = low growth, h = high growth).

nat, character, nationality (ch = Swiss; int = foreign/international).

sex, character (f = female, m = male).

age, numeric, typically ranging from 0 to 100 (incl. >100).

birth_rate, numeric, number of children per year.

births_int_ch, numeric, proportion of children with Swiss nationality born to non-Swiss mothers.

mor, numeric, prospective mortality rate (probability of death).

acq, numeric, rate of acquisition of Swiss citizenship.

emi, numeric, rate of people emigrating abroad.

mig_ch, numeric, national / inter-cantonal net migration (number of immigrants - number of emigrants).

imm_int, numeric, number of people immigrating from abroad.

mig_sub, numeric, within canton net migration. Useful to account for movements between different subregions (e.g., municipalities). This argument is optional.

population

data frame including the starting population of each demographic group. Possible values are the same as in parameters (apart from year).

year character, should be year_first - 1.

spatial_unit character.

nat character.

sex character.

age numeric.

n numeric, number of people per demographic group.

year_first

numeric, first year to be projected.

year_last

numeric, last year to be projected.

age_groups

numeric, number of age classes. Creates a vector with 1-year age classes running from 0 to (age_groups - 1). Defaults to 101 (FSO standard number of age groups).

fert_first

numeric, first year of female fertility. Defaults to 16 (FSO standard value).

fert_last

numeric, last year of female fertility. Defaults to 50 (FSO standard value).

share_born_female

numeric, fraction of female babies. Defaults to 100 / 205 (FSO standard value).

subregional

boolean, TRUE indicates that subregional migration patterns (e.g., movement between municipalities within a canton) are part of the projection.

spatial_unit

character, name of variable containing the names of the region or subregions for which the projection shall be performed.

Value

Returns a data frame that includes the number of people for each demographic group per year (for the starting year and each projected year). The number of rows is the product of all years times all demographic groups (e.g., nationality (2) * sex (2) * age groups (101) = 404). Variables included in the output:

age

integer.

sex

factor, female (f) and male (m).

nat

factor, Swiss (ch) and international / foreign (int).

year

integer, indicating starting year / projected years.

spatial_levels

factor, spatial levels for which the projection was run (e.g., canton, municipalities).

n

double, number of people per demographic group.

Examples

# Run projection for the sample data (whole canton of Aargau)
propop(
  parameters = fso_parameters,
  year_first = 2019,
  year_last = 2022,
  population = fso_population,
  subregional = FALSE
)
#> Running projection for: Aargau
#> ✔ Year: 2019
#> ✔ Year: 2020
#> ✔ Year: 2021
#> ✔ Year: 2022
#> # A tibble: 2,020 × 6
#>      age sex   nat    year spatial_unit     n
#>    <int> <fct> <fct> <int> <fct>        <dbl>
#>  1     0 m     ch     2018 Aargau        2506
#>  2     1 m     ch     2018 Aargau        2589
#>  3     2 m     ch     2018 Aargau        2568
#>  4     3 m     ch     2018 Aargau        2635
#>  5     4 m     ch     2018 Aargau        2710
#>  6     5 m     ch     2018 Aargau        2608
#>  7     6 m     ch     2018 Aargau        2569
#>  8     7 m     ch     2018 Aargau        2611
#>  9     8 m     ch     2018 Aargau        2634
#> 10     9 m     ch     2018 Aargau        2512
#> # ℹ 2,010 more rows